+++ docs/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Next
+
+- Internal: fixed order of implemented ops in OpType enum.
+- StableHLO: added `OptimizationBarrier()` op.
+- For `Value.OpName()`: return "InputParameter" (`stablehlo.InputParameterName`) for input parameters instead of "nil".
+- Added pjrt.Plugin.IsCPU() method.
+
 # v0.2.1: CUDA plugin fixes
 
 - Fixed CUDA library paths -- it was hardcoded to a test directory.